finish project

+++ routes/ideas.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 const Idea = require("../models/Idea");
 
-//Get All ideas
+// Get all ideas
 router.get("/", async (req, res) => {
   try {
     const ideas = await Idea.find();
@@ -13,7 +13,7 @@ router.get("/", async (req, res) => {
   }
 });
 
-//get single idea
+// Get single idea
 router.get("/:id", async (req, res) => {
   try {
     const idea = await Idea.findById(req.params.id);
@@ -24,13 +24,14 @@ router.get("/:id", async (req, res) => {
   }
 });
 
-//Add an idea
+// Add an idea
 router.post("/", async (req, res) => {
   const idea = new Idea({
     text: req.body.text,
     tag: req.body.tag,
     username: req.body.username,
   });
+
   try {
     const savedIdea = await idea.save();
     res.json({ success: true, data: savedIdea });
@@ -40,31 +41,53 @@ router.post("/", async (req, res) => {
   }
 });
 
-//Update Idea
+// Update idea
 router.put("/:id", async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: "You are not authorized to update this resource",
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "Something went wrong" });
   }
 });
 
-//Delete idea
+// Delete idea
 router.delete("/:id", async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // Match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: "You are not authorized to delete this resource",
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "Something went wrong" });
